+++ tests/test_signoff_flow.py
@@ -430,9 +430,9 @@ class SpecificUserGroupsTest(SignoffWebTest, FormattedErrorMixin, unittest.TestC
             cls.source_collection2.replace("alice", "destination"))
 
         settings['signer.group_check_enabled'] = 'false'
-        settings['signer.alice_cid1.group_check_enabled'] = 'true'
-        settings['signer.alice_cid1.editors_group'] = 'editeurs'
-        settings['signer.alice_cid1.reviewers_group'] = 'revoyeurs'
+        settings['signer.alice.cid1.group_check_enabled'] = 'true'
+        settings['signer.alice.cid1.editors_group'] = 'editeurs'
+        settings['signer.alice.cid1.reviewers_group'] = 'revoyeurs'
         return settings
 
     def setUp(self):
@@ -569,9 +569,9 @@ class PerBucketTest(SignoffWebTest, unittest.TestCase):
             cls.destination_bucket])
 
         settings['signer.to_review_enabled'] = 'true'
-        settings['signer.stage_specific.to_review_enabled'] = 'false'
+        settings['signer.stage.specific.to_review_enabled'] = 'false'
 
-        settings['signer.stage_specific.autograph.hawk_id'] = 'for-specific'
+        settings['signer.stage.specific.autograph.hawk_id'] = 'for-specific'
         return settings
 
     def test_destination_and_preview_collections_are_created_and_signed(self):
@@ -660,6 +660,15 @@ class GroupCreationTest(PostgresWebTest, unittest.TestCase):
         self.app.get(self.editors_group, headers=self.headers)
         self.app.get(self.reviewers_group, headers=self.headers)
 
+    def test_groups_are_allowed_to_write_the_source_collection(self):
+        body = {'data': {'members': [self.other_userid]}}
+        self.app.put_json(self.editors_group, body, headers=self.headers)
+
+        self.app.put(self.source_collection, headers=self.headers)
+
+        self.app.post_json(self.source_collection + '/records',
+                           headers=self.other_headers, status=201)
+
     def test_events_are_sent(self):
         patch = mock.patch('kinto_signer.utils.notify_resource_event')
         mocked = patch.start()
@@ -688,6 +697,19 @@ class GroupCreationTest(PostgresWebTest, unittest.TestCase):
         assert r['data']['members'] == [self.userid]
         r = self.app.get(self.reviewers_group, headers=self.headers).json
         assert r['data']['members'] == []
+
+    def test_groups_are_not_touched_if_already_exist(self):
+        resp = self.app.put(self.editors_group, headers=self.headers)
+        editors_timetamp = resp.json['data']['last_modified']
+        resp = self.app.put(self.reviewers_group, headers=self.headers)
+        reviewers_timetamp = resp.json['data']['last_modified']
+
+        self.app.put(self.source_collection, headers=self.headers)
+
+        r = self.app.get(self.editors_group, headers=self.headers).json
+        assert r['data']['last_modified'] == editors_timetamp
+        r = self.app.get(self.reviewers_group, headers=self.headers).json
+        assert r['data']['last_modified'] == reviewers_timetamp
 
     def test_groups_are_not_created_if_not_allowed(self):
         # Allow this other user to create collections.